Add a hook for src commits
Initially copied from the ports hook with ports-specific checks removed.
At some point it would be useful to try and merge common bits, but for
now it's not so much code that the duplication is harmful.

The main checks enabled here are verifying that the second line of the
commit log is blank, and that we're not getting phabricator-style commit
messages.  I ran this script on all src commits from 2025 and verified
that none of the hook failures are false positives.